% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyData.R
\encoding{UTF8}
\name{EasyData}
\alias{EasyData}
\title{Function that returns a dataset}
\usage{
EasyData(dataset, path = NULL)
}
\arguments{
\item{dataset}{Name of data set within the easySdcTable  package}

\item{path}{When non-NULL the data set is read from "path/dataset.RData"}
}
\value{
The dataset
}
\description{
Function that returns a dataset
}
\examples{
 z  <- EasyData("sosialFiktiv")

}
